Resolve #3061: Isolate Passport provider regressions
